moved ExceptionGroup to separate module
